+++ lib/std/decls.nim
@@ -26,7 +26,12 @@ macro byaddr*(sect) =
     typ = def[1]
     ex = def[2]
     addrTyp = if typ.kind == nnkEmpty: typ else: newTree(nnkPtrTy, typ)
-  result = quote do:
-    let tmp: `addrTyp` = addr(`ex`)
-    template `lhs`: untyped = tmp[]
+  when defined(nimHasAliasPragma):
+    result = quote do:
+      let tmp: `addrTyp` = addr(`ex`)
+      template `lhs`: untyped {.alias.} = tmp[]
+  else:
+    result = quote do:
+      let tmp: `addrTyp` = addr(`ex`)
+      template `lhs`: untyped = tmp[]
   result.copyLineInfo(def)